Initial commit of rudimentary working version
I wanted to deploy Nextcloud since a long time but always stayed away
from it, mainly because I find the way most PHP applications are
structured very scary (okay, let me rephrase: I find PHP itself already
very scary), because a lot of projects (including Nextcloud) tend to
write stuff into the document root, which is also (at least in theory)
accessibly from the outside world.

We already have a NixOS module in nixpkgs, but I don't like its
implementation because it makes the configuration file writable to the
PHP FPM user and also allows to install apps via the web interface,
which could allow attackers to execute PHP code on the machine.

The primary goal here is to tear apart Nextcloud to the extent that we
can configure it completely via the NixOS module system, including all
the settings from the individual apps.

Right now, we have an updater which fetches all the apps via the API
from apps.nextcloud.com into a JSON file, which we can use to generate
module options. The implementation doesn't yet try to parse settings, so
at the moment we only have one option called "enable" which is used to
bake in the application into the main derivation.

Also, our settings are currently riddled with XXX and TODO comments, so
while the instance is working and we can upload files, install apps and
whatnot, I wouldn't recommend using this in production at all because a
lot of very important settings are dummies right now and we also don't
have sandboxing yet.

Signed-off-by: aszlig <aszlig@nix.build>