Initialize secrets on first startup
The instanceid and secret value so far were hardcoded, which is not a
very good idea for production systems and as we move towards a state
where we can do our first deployment, it makes sense to generate these
values during initial setup.

This is done via an additional service that's started before the actual
service, which creates /var/lib/nextcloud with user root and group
nextcloud, but only gives execute access to the directory for the
nextcloud group. That way the user can neither list its contents nor
write files inside that directory.

We put a secrets.env file in there that we reference via the main
systemd service and it's accessible by root only. While this still can
be an issue in an attack scenario when the attacker can execute code
it's less of an issue in case of directory traversal vulnerabilities
because the main nextcloud user is not able to read this file. However,
the secrets can still be read via $_ENV.

Signed-off-by: aszlig <aszlig@nix.build>