Sandbox nextcloud and nextcloud-cron
In the long term I want to also prevent Nextcloud from binding to
sockets, but I first need to figure out whether we can really do that or
if there are legitimate uses for that. Ideally we'd just use a network
namespace, but there are already a plethora of occurences where
Nextcloud communicates to the outside world.

Nevertheless however, we now have at least a chroot for the main service
and the cron service, which only allows access to the required store
paths and the data directory.

Unfortunately, there are also invocations of shell_exec() in Nextcloud
and some apps, so we need to provide a /bin/sh inside the sandbox.

I also changed the libreoffice path to use the unwrapped variant, so we
don't start up dbus and all the stuff we really don't need to generate
previews.

Signed-off-by: aszlig <aszlig@nix.build>