Generate info about shipped apps as well
So far, we have only generated applications from the application store
at https://apps.nextcloud.com/, however users are able to disable the
applications that come shipped with Nextcloud as well, although it's not
possible to uninstall them.

For our upcoming implementation for enabling/disabling applications in a
declarative way using NixOS options, this is quite handy to have.

Signed-off-by: aszlig <aszlig@nix.build>