Implement app enable/disable via NixOS options
First of all, we now disable all the apps during the nextcloud-init
derivation, so that applications we don't want don't increase the size
of the closure and don't end up on a newly deployed system.

Second, we enable/disable apps during a new oneshot-service prior to the
actual Nextcloud server, which now also handles upgrades. This should
allow us to enable and disable apps directly with NixOS options and we
can soon remove that feature from the web UI.

This still has one issue though, when you have an enabled app that's
enabled for everybody and want it to be enabled only for certain groups,
it won't work (and won't work vice versa).

Signed-off-by: aszlig <aszlig@nix.build>