Add implementation for LibreOffice Online
I originally started packaging OnlyOffice, but getting this into a state
where I would be satisfied would require a lot of more work than what we
have now.

While LibreOffice Online runs on the server and thus is more
memory-hungry I nevertheless prefer the architecture to OnlyOffice, as
it has a strong focus on self-sandboxing.

However, we still needed to add a few patches to bend LibreOffice Online
into our NixOS configuration with a focus an bringing the already
existing sandboxing solution to the next level.

There are still a few quirks (like eg. inserting diagrams causes a mutex
deadlock) and things we need to improve, but so far it's working fine.

Our enhancements to sandboxing are first of all using network namespaces
to only allow to the local web server from the LibreOffice Online
instance. We do that via two proxies, one being the internal proxy that
forwards everything to a Unix socket, which in turn gets picked up by
the external proxy (simply an instance of systemd-socket-proxyd) that
connects back to nginx. This obviously could be better isolated, so that
the LibreOffice Kit only has access to the Nextcloud locations on our
webserver, but that's currently low-priority and less of a problem if
you're running a Nextcloud instance with trusted users.

Another very ugly thing right now is that the websocket server has
CAP_SYS_ADMIN capabilities, which provides quite a lot of syscalls to
the main process (they will be dropped however). Bringing down that
attack surface is also an area where we'd need some work.

Signed-off-by: aszlig <aszlig@nix.build>