lool: Directly patch in discovery/capabilities
LibreOffice Online doesn't give a list of capabilities, so we can simply
patch in an empty array whenever needed.

For the discovery, this is a (little) bit more involved because the URL
of loleaflet.html contains a hash, which I had to rearrange a bit so we
can look it up more easily from the base package.

With a small Python helper, we simply extract all the mime types
LibreOffice Online supports and directly hardcode the loleaflet URL in
the source of the richdocuments application.

The reason I'm doing this is to avoid problems if we run Nextcloud in a
development VM which uses a different HTTP(S) port in the guest. Let's
say we have port 8000 (which then also is the port used for the
websocket) and the nginx is running at port 80 in the VM. Now you can
choose between either using port 8000 or 80 and either the discovery or
the websocket connection will fail, except you do some port redirection
magic of course.

Another advantage of this is that we now have less moving parts, because
sometimes it was the case that the cached discovery.xml file from
richdocuments was occasionaly out of date. By hardcoding the information
in the source code this can't happen anymore.

Signed-off-by: aszlig <aszlig@nix.build>