lool: Simplify the proxy chain
This gets rid of the socat/systemd-socket-proxyd services and instead
uses ip2unix to let LOOL connect to an nginx socket that represents the
Nextcloud virtual host.

Using nginx directly to provide the socket has another advantage because
it limits the virtual hosts that are available to LOOL to just the
Nextcloud domains.

Signed-off-by: aszlig <aszlig@nix.build>