lool: Implement support for fontconfig
So far we only had two shipped fonts, which is not a very good selection
for pre-existing documents. Instead, let's use a very opinionated list
of fonts to use as default fonts. Note, that the Microsoft Core Fonts
are not included in that list because they're non-free.

With this, we no longer need any system templatm anymore for the chroot,
so removing that is comping up next :-)

Signed-off-by: aszlig <aszlig@nix.build>