lool: Remove system templates
These are only needed on distros that do not use Nix for package
management. With Nix however, we just use the individual store paths
that we need for the chroot and don't need additional stuff. If we still
want additional packages, we can simply introduce them via the Nix
store.

Signed-off-by: aszlig <aszlig@nix.build>