Fix disabling/enabling apps during upgrade
If we have a new Nextcloud installation, everything is fine because we
only need to enable apps that we haven't enabled so far.

However, if we already have an existing Nextcloud installation and later
want to enable a specific app, the operation will fail, because enabling
apps is done on the basis of the *old* Nextcloud store path, which
doesn't contain the app.

Now, we only disable apps using the old store path and enable them using
the new store path.

Signed-off-by: aszlig <aszlig@nix.build>