lool: Make logging less verbose
Even when setting a lower log level than trace, LibreOffice Online sets
"trace" during startup of loolwsd and ForKit. I find this to be a
misfeature making logs still hard to dig through, so I removed it and
the logLevel option now sets it globally, not only after startup.

Another thing I've changed is the "Waking up dead poll thread" warning,
which frequently occurs, to use the "information" log level because
that's IMHO where it should belong as it's not cruicial for running
loolwsd.

Signed-off-by: aszlig <aszlig@nix.build>