lool: Add support for TLS
Ideally all communication between LibreOffice Online and Nextcloud would
be unencrypted, because it's low-overhead and we are only using Unix
domain sockets anyway.

So far I've only done tests without TLS, so when using TLS for the first
time, the LibreOffice Online server stopped working because first of all
it gave the client a "ws://" instead of a "wss://" URL and second
because the connection back to the web server was done also done via TLS
and it wanted to verify the certificate.

The latter really isn't necessary as we are using a Unix socket, so I
set ssl.enable to false but ssl.termination to true, so that loolwsd
sends the appropriate wss:// URL and also doesn't check certs when
connecting back to Nextcloud.

Along the way, I also cleaned up the settings a bit, for example
specifying the server_name isn't necessary anymore.

Signed-off-by: aszlig <aszlig@nix.build>