lool: Use discovery XML file from source
Referencing the discovery.xml of the resulting build of LibreOffice
Online is fine, but it has one major drawback: Whenever we change
something in the LibreOffice Online implementation, Nextcloud itself is
also rebuilt.

Using the XML file from the "src" attribute no longer causes such a
rebuild and only rebuilds if we have a different source.

Signed-off-by: aszlig <aszlig@nix.build>