Fix and improve updating of Nextcloud and apps
So far we have used a GC root to memorize the old version and config in
order to have a reference for the updater to figure out what the base
version is.

However, we really only need the "version" configuration value, so
instead of a GC root, we now have a file at /var/lib/nextcloud/.version
which tracks the previous version (and just the version, nothing more).

The version is then brought into the config via __NEXTCLOUD_VERSION,
which by default is the current version and just the update process gets
the value from /var/lib/nextcloud/.version.

In addition to that, I also added a small patch for Nextcloud so that it
just silently discards changes to the config file and will continue
regardless of whether the config file is writable or not.

On most systems, Nextcloud is still running while the upgrade takes
place, so the main reason (among setting the version and other values)
the updater wants to write to the config is to enable maintenance mode
so the site becomes inaccessible during an upgrade.

Fortunately we don't need to do such things, because our uWSGI server is
down during upgrade and the cron service also waits for the upgrade to
complete. We also use socket activation for uWSGI, so incoming clients
will just wait until the upgrade is finished.

I have tested this by upgrading from Nextcloud 15.0.3.1 to 15.0.4.0 plus
a few apps alongside.

Signed-off-by: aszlig <aszlig@nix.build>