Allow setting app-specific configuration options
Right now, this is only done while an app is going to be enabled, but
it's easy to add another state file where we can diff the old settings
with the new ones and update them accordingly.

We could also just diff the settings with the ones that are already in
the database, but for that we first need to improve on performance of
the OCC tool. Right now it takes around 4 to 5 seconds in my test VM to
execute a single OCC command, which is more than subpar.

I also flagged the new config option as internal, because its values are
not necessarily strings but could be anything depending on the app. A
lot of apps use JSON for some settings, but I wouldn't bet on "either
string or JSON".

Signed-off-by: aszlig <aszlig@nix.build>