lool: Use more strict nginx location config
First of all, let nginx take care about most of the static files,
because the only exception to those files is loleaflet.html which is
generated based on a template by the loolwsd file server implementation.

Second, let's actually remove all the URLs we don't really need, like
for example the discovery/capabilities URLs, which we have made obsolete
a while ago (the information is directly patched into the richdocuments
app).

Signed-off-by: aszlig <aszlig@nix.build>