updater: Fix retrieving Nextcloud download URL
So far we have hardcoded the base URL for the download and just appended
the corresponding version to the path. While I could have fixed this by
truncating the version to 3 digits, I didn't do so, because the update
server already provides an URL to the ZIP file.

In our case, we want to have the tar.bz2 instead, so we now use the ZIP
file URL and just replace the extension with .tar.bz2.

Signed-off-by: aszlig <aszlig@nix.build>