Add onlyoffice WIP tree
I decided to abandon OnlyOffice (yes, I refuse to write it all upper-
case) in favour of LibreOffice Online, which not only has a way better
code quality but also a better privilege separation model.

I'm adding this only[TM] for reference and for people who might want to
pick up on this.

The whole subtree is currently very unstructured, so it also will need a
cleanup at some point as well, should someone want to pick up on that
work.

Signed-off-by: aszlig <aszlig@nix.build>