Generate rewrite/redirect locations
This is mainly so we don't need to spell out everything twice and don't
need to use a regex, because we actually only allow two variants for
them:

One without the slash at the end and another one with the slash.

So instead of copy & pasting all that stuff, let's just generate this on
the Nix expression level.

Along the way this also fixes the ocm/ocs-provider URLs so that no
/index.php needs to be appended.

Signed-off-by: aszlig <aszlig@nix.build>