xmpp: Introduce MongooseIM 3.2.0 base package
This should get us started with all the Erlang dependencies that are
locked from upstream. Currently, we don't have something like rebar2nix
available, so I've built something like "rebar2nix lite", which is a
mixture of Erlang and Python code.

The reason why I used Python is because there is no JSON library in the
Erlang default distribution and at that stage, we want to run with very
minimal dependencies.

Unfortunately the major work involved here was getting around the
buildRebar3 implementation in nixpkgs, which implicitly expects that
every package has to be available on hex.pm, because it uses a static
Hex registry snapshot.

To get around that, I've written a small helper called fix-registry.erl,
which essentially throws away the pinned snapshot and generates a new
one based on the symlinked packages.

Another thing we do is "jailbreaking" (I used this term from the Haskell
packaging in nixpkgs) all dependencies, that is, getting rid of all the
version constraints and options around that (eg. "{pkg, somename}").

We already gathere all the versions using update.py, so there is really
no need to pin dependencies on the package build level.

So what's left now is figuring out to deceive relx so that it doesn't
want to write stuff all over the place or - alternatively - get rid of
relx entirely and start BEAM on our own. We shall see...

Signed-off-by: aszlig <aszlig@nix.build>