xmpp: Use relx for generating MongooseIM package
I opted to use a more conservative approach that what I did on the
Headcounter deployment, where I teared apart a lot of MongooseIM[1] in
order to create NixOS options of almost every setting.

Instead of doing most things on the Nix side, we now generate the full
release using the "release" sub-command of rebar3. This of course also
gets us a bunch of shell scripts, which isn't that great (as I'd
personally want more control on that) but at least it doesn't break on
every update of MongooseIM and thus is more maintainable.

However, I still added a few patches, so that we can better integrate it
into Nix(OS) and systemd, eg. by logging to stdout or changing the
output paths accordingly to fit Nix store path convention.

[1]: https://github.com/headcounter/deployment/tree/c74d9b63a4aa7c0c/modules/services/mongooseim

Signed-off-by: aszlig <aszlig@nix.build>