lool: Restrict permissions on internal socket
Nginx creates the socket with mode 0666, which makes the socket
available to all users on the system.

We can't easily change that mode without patching nginx, so let's
instead create the socket in a subdirectory where the directory
permissions are restrictive enough so that only the LibreOffice Online
user (and the nginx user of course) has permission to access it.

Signed-off-by: aszlig <aszlig@nix.build>