xmpp: Add authentication for Nextcloud
This is done via a custom MongooseIM authentication module instead of
using xmpp-cloud-auth[1], because MongooseIM already has an HTTP client
using FUSCO and executing another command will make it harder for us to
sandbox and we'd also need one more service.

Another derivation from the xmpp-cloud-auth way is that we solely rely
on Unix domain sockets, so not only don't we need to take care of open
TCP sockets but we can also avoid needing credentials for authenticating
against Nextcloud.

We simply differentiate whether the auth request is coming from the web
or whether it's coming from the internal Unix socket. I implemented this
by adding another uWSGI parameter to the mix, which will only be set for
the Unix domain socket. Using such a parameter instead of an HTTP header
has the advantage that there is no way to forge it from outside.

[1]: https://github.com/jsxc/xmpp-cloud-auth

Signed-off-by: aszlig <aszlig@nix.build>