Move chroot stuff into a dedicated generic module
My intention is to upstream this into nixpkgs very soon, because I tend
to do this fairly often for a bunch of services and it's getting tedious
and also very verbose and the disconnect between systemd.packages and
the actual service configuration makes the NixOS modules harder to read.

So having this as a pluggable NixOS module makes it way more reusable
even when it's not yet upstream and we're going to also apply this to
the XMPP config very soon, so we'll safe a few lines.

Signed-off-by: aszlig <aszlig@nix.build>