xmpp: Disable S2S for now
This is done by putting MongooseIM into a network namespace because
everything except S2S is done via Unix domain sockets.

I'm disabling this for now, because I haven't found a good way to re-use
the TLS certificate and key from nginx in a way that works well in a
common configuration.

Signed-off-by: aszlig <aszlig@nix.build>