Always set app configuration if it has changed
So far we only set the config whenever we have enabled an app, which is
fine in theory but whenever we add new features to this configuration
the deployed apps don't get their configuration updated.

New, we're always setting the new app configuration but only if they
have changed. This is because occ is very slow, so we want to run as few
commands as possible.

Signed-off-by: aszlig <aszlig@nix.build>