xmpp: Set app config for ojsxc
So far we only set up an XMPP server whenever the app is enabled, but
the config is missing and had to be set by an admin.

With this, the values are automatically set by our updater.

Note that I'm setting the apiSecret here to "not needed" because we have
our own authentication module directly implemented in MongooseIM.

Signed-off-by: aszlig <aszlig@nix.build>