xmpp: Fix dependency updater and GitHub URLs
During development of the updater, I apparently got the right GitHub
archive URLs, but later broke it accidentally.

I didn't notice because the sources were already fetched and won't be
tried to be fetched again as long as the hashes are correct.

So when building this on another machine, fetching the sources has
failed for obvious reasons.

The bug in the updater was that I used two variables to refer to the
owner/repo pairs but I only stripped the .git on the combined variant
that's used to fetch the source in advance but not the individual parts
that end up in upstream.json.

Signed-off-by: aszlig <aszlig@nix.build>