lool: Pin LibreOffice to a working nixpkgs version
When deploying to a system with NixOS 18.03, LibreOffice is a bit older
and doesn't build with LibreOffice Online, so in order to make sure we
have a version that's definitely working, let's pin nixpkgs to a version
that I have tested before.

In the long term I think we should replace this by a full source build
of the SDK, which only compiles the necessary parts we need and it also
matches the LibreOffice Online version. The latter is not the case at
the moment, which I believe could be the reason why certain actions
don't work. So this would be another reason for going with a full
version-specific source build.

Signed-off-by: aszlig <aszlig@nix.build>