xmpp: Forge remote addr to 127.0.0.1 for auth
The internal nginx virtual host that listens to a Unix domain socket
doesn't work if brute force protection is enabled, because it tries to
parse the Unix socket address as an IP address, which doesn't work.

We don't allow external connections and the socket path is only
accessibly by the MongooseIM user, so forcing REMOTE_ADDR to 127.0.0.1
is safe here.

I didn't notice this so far because during development I disabled brute
force protection.

Signed-off-by: aszlig <aszlig@nix.build>