Fix temporary directory
The Nextcloud instance is chrooted and /tmp is a tmpfs there, so
uploading files could be problematic on systems with a low amount of
memory available (or left).

So instead, let's just use the CacheDirectory service option and use
that directory in the Nextcloud config and also as PHP's upload_tmp_dir,
so that we also cover apps that don't use Nextcloud's TempManager class.

Signed-off-by: aszlig <aszlig@nix.build>