systemd-chroot: Fix type for chroot.packages
When using types.package as the first type that is matched, the
corresponding value is passed to toString, which even if it's a string
with context and starts with a slash will be converted to a path.

This unfortunately will also try to convert invalid paths and things
such as "${somederivation} foo" into "/nix/store/...somederivatien foo",
which isn't a valid path as it contains spaces.

Reversing the types used for either avoids this, because a string with
context will have the typecheck for types.str to succeed and thus the
types.package check won't run at all.

Signed-off-by: aszlig <aszlig@nix.build>