Set session.save_path to something other than /tmp
By default, the value of session.save_path is /tmp, which is not a very
good solution because it's in a private /tmp.

While Nextcloud uses its own session handler, some data is still saved
in PHP sessions, so while at it let's actually save it in
/var/cache/nextcloud/sessions and also change the temporary upload
directory to /var/cache/nextcloud/uploads.

Using both of these paths in the initiall install derivation is adding
some failures during that to the mix, so I splitted the PHP common into
commonPhpConfig (which we had before) and runtimePhpConfig, the latter
only being executed outside of Nix builds.

Signed-off-by: aszlig <aszlig@nix.build>