Fix /ocs-provider URL
Apparently, I fucked this up and when I introduced this route in
b80778d896d6125eba763e29b73b46bfce5a97ac, I simply replaced "some
character" with an "s".

Unfortunately however, it was the wrong "s", so instead of
"ocs-provider" we suddenly got "osm-provider".

Thanks to @shinice for helping in discovering this.

Signed-off-by: aszlig <aszlig@nix.build>