nextcloud-init: Don't enable encryption by default
First of all, the nextcloud-init service doesn't run in a sandbox, so we
really don't want this service to execute OCC commands.

Second, the OCC command actually fails, because it can't find
__NEXTCLOUD_VERSION, which got introduced way later than when I
introduced running encryption:enable command.

If we really want to have something like that (that is, automatically
encrypting files), we can do so in a separate module, but for now let's
remove it.

Signed-off-by: aszlig <aszlig@nix.build>