Fix writing to /var/lib/nextcloud/.version
On a new deployment, there is no .version yet, so bind-mounting that
path will fail. In order to cope with this, I'm now bind-mounting the
whole /var/lib/nextcloud path.

While this might sound like it would introduce some critical attack
vector, it isn't really that critical because we still run the main
process as a different user. Even within the user namespace the user of
the main process shouldn't have access to the .version file.

So in the event that for some reason the user of the main process is
able to write to .version, it's still absolutely non-critical because
the only thing that's possible is breaking its own update procedure.

Signed-off-by: aszlig <aszlig@nix.build>