Add a test for Nextcloud upgrades
This will essentially try to update Nextcloud from the oldest version we
have shipped here (15.0.4) to the latest one. In order to do so, we
introduce a new directory in package/old, which contains the Nix
expression plus patches for 15.0.4.

Right now there are not many differences and the patches still apply for
both the current and the old version, but I'm keeping that copy anyway
in order to avoid that we accidentally update the base version for the
upgrade.

The test itself is pretty straightforward with two system generations
going from the old base version and switching to the new version. At the
end we also check whether the apps we want to have enabled are really
enabled, just to be sure.

Another thing to note here is that instead of a normal let binding, we
now have an internal nextcloud.package option, which we also use within
the test to override the package with the old one.

Signed-off-by: aszlig <aszlig@nix.build>