Set CA root certificate location in more places
It's a tad ugly to set several environment variables to deal with this,
especially since there are occasions where SSL_CERT_FILE and
NIX_SSL_CERT_FILE are accepted.

So I'm setting both variants so we're absolutely sure we get the right
one in all cases and I also set openssl.cafile for the PHP config.

This should hopefully get rid of most (if not all) of the certificate
verification errors.

Signed-off-by: aszlig <aszlig@nix.build>