Fix value for overwrite.cli.url
This needs to be an absolute URL and a few applications determine the
right domain name out of this value.

I have initially marked this with XXX, but removed the comment because
everything[TM] seemed to work. However this is not the case and it's
used for retrieving absolute URLs within the CLI.

Signed-off-by: aszlig <aszlig@nix.build>