Fix nginx URL configuration and rewrites
First of all, our rewrites so far weren't quite right, because what
Apache does is setting the REQUEST_URI to the *destination* of the
rewrite, which is also what the enpoint expects. So we now set the
REQUEST_URI accordingly to the destination URI instead of the original.

This fixes all of the rewrites we had so far and while at it, I also
removed the rewrites for host-meta and host-meta.json, because they're
unused and are only handled by the no longer maintaned user_webfinger[1]
app.

Another thing that I'm removing are the error handlers for HTTP 403 and
404, because they break 40[34] error codes with JSON and instead
redirect to the login page.

I've added another NixOS VM test to make sure the URL config doesn't
break in the future, should I fuck it up again.

[1]: https://github.com/nextcloud/apps/tree/5e525f8ccbad79/user_webfinger

Signed-off-by: aszlig <aszlig@nix.build>