updater: Sanitize download file name
The "spgverein" app has a "?" in its download file name, which in turn
causes the nix-prefetch-url call to fail. So let's strip off the
question mark and everything after it and also remove characters that
are otherwise dangerous, like eg. dots in front of the filename. The
latter shouldn't be a problem though, because directory traversal would
only be possible if slashes would also be allowed, but let's do it
anyway so we don't run into errors in the future.

Signed-off-by: aszlig <aszlig@nix.build>