uwsgi: Disable timeout for response body
Ideally the timeout would only occur if receiving the response body has
completely stalled and is idle for $timeout seconds.

This could potentially make us DoSable (by just stalling the connection
long enough) though, so let's test this in production and fix it
properly later X-D

The issue this works around is that the connection aborts whenever you
have a large download, so the connection is not idle but the timeout
still occurs.

Signed-off-by: aszlig <aszlig@nix.build>