Revert "uwsgi: Disable timeout for response body"
This reverts commit 1c16b2baad427f3f4eb96839a831e4df56c0592e.

As mentioned in that commit, this makes us DoSable and after
experimenting around in a (not that important) production system, I
decided it would be better to fix this properly when we have a proper
NixOS VM test.

The problem actually only exists for really large files, so we should
rarely bump into this.

Signed-off-by: aszlig <aszlig@nix.build>