Add coturn configuration for Nextcloud Talk
While there is already a default STUN server pointing to
stun.nextcloud.com:443, there isn't a default one for TURN.

The reasons for this are pretty obvious, because TURN is for relaying
the traffic if a peer connection isn't possible.

This adds a STUN and TURN server called coturn, which of course is
sandboxed in the same way as we do for other components. Unfortunately,
we can't simply re-use the existing SSL certificate and key of nginx, so
I added a small helper service, which copies over the cert and key with
the right permissions.

I also added two small patches for allowing to use environment variables
to pass around the secrets. Ideally we'd go completely without these
secrets, but unfortunately I haven't found a very nice solution for that
yet.

Signed-off-by: aszlig <aszlig@nix.build>