coturn: Make backwards-compatible with NixOS 18.09
First of all, NixOS 18.09 didn't have the types.port alias, so let's
actually use types.ints.u16 instead.

Another thing that's different in 18.09 is that we have an older coturn
version there, so I added a small version comparison to use a more
recent version if the version of the current nixpkgs is too old.

Signed-off-by: aszlig <aszlig@nix.build>