coturn: Work around ACME cert/key locations
If ACME is used for an nginx virtual host configuration, the values for
sslCertificate and sslCertificateKey are undefined and they're only
assigned as local attributes within the nginx module.

So in order to refer to the correct location, we pretty much need to
replicate what is being done in the nginx module.

Signed-off-by: aszlig <aszlig@nix.build>