coturn: Log everything to stdout
While coturn is running in a chrooted tmpfs, it still is able to write
to the tmpfs and thus to turn.log. The code has several fallbacks and
only uses stdout as a last resort.

Having such a file only available when accessing /proc/N/fd/X is not
very convenient, so let's actually make sure that turn.log isn't created
in the first place.

Signed-off-by: aszlig <aszlig@nix.build>