lool: Fix nginx configuration
With newer nixpkgs, gixy is used for checking the nginx configuration
and with LibreOffice Online enabled, the build will fail.

However, the error about the alias directive is a false-positive and
only is a problem whenever *either* the location *or* the alias
directive contain a slash at the end. If both either end with slashes or
both don't end with slashes, it's not a problem.

So in order to shut up gixy, let's just append a slash to the location
and the alias directive.

Another gixy error was on using $http_host, but this error also isn't a
real issue, because it is used on a dedicated Unix domain socket that is
only accessibly by the libreoffice-online user.

Signed-off-by: aszlig <aszlig@nix.build>