lool: Explicitly provide server name
When running Nextcloud configured with a port other than 80 or 443, the
leaflet is trying to connect to the websocket at the host without
specifying the port.

Since the websocket URL is comprised of either ws:// or wss:// followed
by the server name and then followed by the path, we have to embed the
colon and port inside the server name, hence we set the server_name
option directly to the configured host with an optional port.

Signed-off-by: aszlig <aszlig@nix.build>