updater: Factor out common types into own module
This is mainly so we can seperate out all this mess that I've written in
a very quick & dirty way because I initially thought it would be "just a
small updater".

In the end, we want to have two instances of nested data structures,
which represent the current state and the new state. By diffing both, we
then can download the appropriate packages we need and have pure data
structures to work with instead of mutating the existing ones.

Signed-off-by: aszlig <aszlig@nix.build>