updater: Add missing type stubs
This adds stubs for defusedxml, pyopenssl, semantic_version and tqdm
only for the stuff we actually use.

Alongside typing them, I converted some of the integer constants to
newtypes so we get a type error if we either use the wrong constant or
if we use integer literals directly.

I originally only wanted to provide types for semantic_version, because
that's the most cruicial one because we use Version objects in several
places, but while at it I thought it would be a good idea to get rid of
all those ignored imports.

The defusedxml types were probably the easiest to implement because we
only need to re-export the attributes of the ElementTree module from the
standad library.

Signed-off-by: aszlig <aszlig@nix.build>