updater: Factor out functionality from main.py
While I'm still not satisfied with the naming of the individual modules
and the locations of the functions, we now have ReleaseInfo as the
primary data structure that contains all the information about Nextcloud
Server and its applications.

We now have a new ReleaseDiff class, which is a small helper to query
and join the differences between two ReleaseInfo structures.

The individual apps now also have different types, InternalApp for apps
shipped with Nextcloud Server and ExternalApp for apps available at the
official app index.

With this, we now have moved most of the functionality out of main.py
and also have a way to better work with the ReleaseInfo data structures
so we can implement fetching multiple upstream major versions separately
from each other.

In order to facilitate testing, I currently hardcoded the version we
send to the Nextcloud updater API to version 14, because otherwise we'd
get version 16, which we do not want yet.

Signed-off-by: aszlig <aszlig@nix.build>