updater: Use dirindex to get Nextcloud versions
The directory index of https://download.nextcloud.com/server/releases/
lists *all* of the releases instead of picking one for us based on
installation time, which the updater API does.

We simply use BeautifulSoup to carve out all the links of the directory
index and match it against a simple regex. This also has the advantage
that we don't need to parse any XML API response nor do we need to alter
the file suffix from zip to tar.bz2.

I also thought I could simplify the version specification and we no
longer need the NextcloudVersion type, because the release tarballs use
3 digit version numbers.

Unfortunately, the fourth digit is also relevant when it comes to
database upgrades, so for the time being, we need to keep it. However,
instead of the NextcloudVersion type, we now use the build field of
semantic_version.Version to encode the fourth digit.

Signed-off-by: aszlig <aszlig@nix.build>