Rename package/current to packages/15
We're going to have several directories named after the corresponding
Nextcloud major version, so we no longer have just old and current.

The old directory still stays the same, because it's used as a fixture
for the upgrade VM test, but everything in that directory that's solely
comprised of digits will soon be recognized as a Nextcloud major
version.

Signed-off-by: aszlig <aszlig@nix.build>