updater: Add ability to fetch all major versions
We now loop through all the directories in packages/ which solely
consist of digits and update upstream.json for each directory, fetching
the specifications according to the major version equal to the directory
name.

The implementation currently is pretty dumb and doesn't do any caching,
so it fetches all the app packages again and again, but this only should
be an issue when fetching apps for the first time or when a whole bunch
of applications are updated/added upstream.

This is the first step towards supporting multiple major versions and
what we need to do now is actually add Nextcloud 16 and some module code
to distinguish between major versions and app collection.

Signed-off-by: aszlig <aszlig@nix.build>