nc15/files_rightclick: Exclude version 0.15.[01]
This has been bugging me for at least three updates, where I always
manually needed to remove the updates for files_rightclick.

With this change in the updater, we now have a new attribute called
"constraints", which we can use in upstream.json to further constrain
the version selection for a particular app.

Of course, the files_rightclick is the first of such constraints to add,
banning version 0.15.0 and 0.15.1 from being updated in the hopes that
either 0.15.2 or 0.16.0 will fix the issue.

Signed-off-by: aszlig <aszlig@nix.build>