updater: Don't fetch apps twice
Ideally, we would just gather a set of apps that we need to download and
decide on that set the list of the things we need to download, but to
make it at least a bit less painful, we're now caching the downloads in
fetch_app_hash().

We also now no longer write the upstream info files after updating each
Nextcloud major version in preparation for the ideal case described
above.

Signed-off-by: aszlig <aszlig@nix.build>