packages/updater: Move themes into upstream.info
Right now we only have the breeze-dark theme, but when fetching several
Nextcloud major versions it becomes quite tedious to keep the themes up
to date, even if we just have one.

Luckily, @mwalbeck uses branches for the specific Nextcloud major
versions, so we don't need to do complex inspection of the Git
repository to find out the Nextcloud major version for a given source
tree.

I've accordingly improved the updater and moved the theme from the main
package expression into upstream.json.

Signed-off-by: aszlig <aszlig@nix.build>